[storage-proto] move TreeState and StartupInfo to storage-interface
