+++ execution/executor-types/src/lib.rs
@@ -17,7 +17,7 @@ use libra_types::{
 };
 use scratchpad::{ProofRead, SparseMerkleTree};
 use std::{cmp::max, collections::HashMap, sync::Arc};
-use storage_proto::TreeState;
+use storage_interface::TreeState;
 
 /// A structure that summarizes the result of the execution needed for consensus to agree on.
 /// The execution is responsible for generating the ID of the new state, which is returned in the
