+++ execution/executor/Cargo.toml
@@ -29,7 +29,6 @@ libra-vm = { path = "../../language/libra-vm", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 scratchpad = { path = "../../storage/scratchpad", version = "0.1.0" }
 storage-interface = { path = "../../storage/storage-interface", version = "0.1.0" }
-storage-proto = { path = "../../storage/storage-proto", version = "0.1.0" }
 
 [dev-dependencies]
 proptest = "0.9.6"
