+++ secure/json-rpc/Cargo.toml
@@ -32,5 +32,4 @@ libra-temppath = { path = "../../common/temppath", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0", features = ["fuzzing"] }
 libradb = { path = "../../storage/libradb", version = "0.1.0" }
 storage-interface = { path = "../../storage/storage-interface", version = "0.1.0" }
-storage-proto = { path = "../../storage/storage-proto", version = "0.1.0" }
 vm-validator = { path = "../../vm-validator", version = "0.1.0" }
