+++ storage/storage-proto/Cargo.toml
@@ -20,6 +20,7 @@ serde = { version = "1.0.106", default-features = false }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
+storage-interface = { path = "../storage-interface", version = "0.1.0" }
 
 [build-dependencies]
 tonic-build = "0.2"
